LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/security-window/window-security.sub-expected.txt

Blocked access to external URL http://www1.localhost:8800/
CONSOLE MESSAGE: line 16: ReferenceError: Can't find variable: fr_load

Harness Error (FAIL), message = ReferenceError: Can't find variable: fr_load

NOTRUN Window Security testing 
